Authorization and Authentication Complete

+++ Controllers/DogController.cs
@@ -4,6 +4,8 @@ using Microsoft.AspNetCore.Mvc;
 using DogGo.Models;
 using System.Collections.Generic;
 using System;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 
 namespace DoGo.Controllers
 {
@@ -19,15 +21,19 @@ namespace DoGo.Controllers
         // GET: DogController
         public ActionResult Index()
         {
-            List<Dog> dogs = _dogRepo.GetAllDogs();
+            int ownerId = GetCurrentUserId();
 
-            return View(dogs);
+            List<Dog> dog = _dogRepo.GetDogsByOwnerId(ownerId);
+
+            return View(dog);
         }
 
         // GET: DogController/Details/5
         public ActionResult Details(int id)
         {
-            return View();
+            Dog dog = _dogRepo.GetDogById(id);
+
+            return View(dog);
         }
 
         // GET: DogController/Create
@@ -37,12 +43,16 @@ namespace DoGo.Controllers
         }
 
         // POST: DogController/Create
+        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create(Dog dog)
         {
             try
             {
+                // update the dogs OwnerId to the current user's Id
+                dog.OwnerId = GetCurrentUserId();
+
                 _dogRepo.AddDog(dog);
                 return RedirectToAction("Index");
             }
@@ -55,13 +65,23 @@ namespace DoGo.Controllers
         // GET: DogController/Edit/5
         public ActionResult Edit(int id)
         {
+
+            int ownerId = GetCurrentUserId();
+
             Dog dog = _dogRepo.GetDogById(id);
             if (dog == null)
             {
                 return NotFound();
             };
-
-            return View(dog);
+            if (ownerId == dog.OwnerId)
+            {
+                return View(dog);
+            }
+            else
+            {
+                return NotFound();
+            }
+            
         }
 
         // POST: DogController/Edit/5
@@ -69,23 +89,38 @@ namespace DoGo.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(int id, Dog dog)
         {
-            try
+
+            dog.OwnerId = GetCurrentUserId();
+
+            if(dog.OwnerId == dog.OwnerId)
             {
                 _dogRepo.UpdateDog(dog);
                 return RedirectToAction("Index");
             }
-            catch(Exception ex)
+            else
             {
-                return View(dog);
+                return NotFound();
             }
         }
 
         // GET: DogController/Delete/5
         public ActionResult Delete(int id)
         {
-            Dog dog = _dogRepo.GetDogById(id);
+            int ownerId = GetCurrentUserId();
 
-            return View(dog);
+            Dog dog = _dogRepo.GetDogById(id);
+            if (dog == null)
+            {
+                return NotFound();
+            };
+            if (ownerId == dog.OwnerId)
+            {
+                return View(dog);
+            }
+            else
+            {
+                return NotFound();
+            }
         }
 
         // POST: DogController/Delete/5
@@ -93,15 +128,23 @@ namespace DoGo.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Delete(int id, Dog dog)
         {
-            try
+            dog.OwnerId = GetCurrentUserId();
+
+            if (dog.OwnerId == dog.OwnerId)
             {
-                _dogRepo.DeleteDog(id);
+                    _dogRepo.DeleteDog(id);
                 return RedirectToAction("Index");
             }
-            catch (Exception ex)
+            else
             {
-                return View(dog);
+                return NotFound();
             }
+        }
+
+        private int GetCurrentUserId()
+        {
+            string id = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return int.Parse(id);
         }
     }
 }
